Add rating functionality to Movie class
Introduced a rating system in the Movie class, allowing multiple ratings to be added and averaged. Updated Main.java to demonstrate adding ratings and printing updated movie info. Added MovieManager.java as a new class for managing a list of movies. Fixed a bug in getLength() and improved Movie's toString formatting.